% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{MPIHelloWorld-package}
\alias{MPIHelloWorld-package}
\title{MPI_HelloWorld}
\description{
Hello World with MPI
\tabular{ll}{
Package: \tab MPI_HelloWorld\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2018-06-12\cr
License: \tab LGPL\cr
}
}
\author{
Tom Sherman, Tiger Gao
}

